Extract hook request resource resolution module
